refactor: share sync utilities
